# pmp.client/src/api/generated/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "pmp.Server API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Cards": {
      "get": {
        "tags": [
          "Cards"
        ],
        "operationId": "Cards_GetCards",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardSummaryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cards"
        ],
        "operationId": "Cards_CreateCard",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardSaveRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cards/{id}": {
      "get": {
        "tags": [
          "Cards"
        ],
        "operationId": "Cards_GetCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetailResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cards"
        ],
        "operationId": "Cards_UpdateCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardSaveRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cards/{id}/archive": {
      "post": {
        "tags": [
          "Cards"
        ],
        "operationId": "Cards_ArchiveCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardSummaryResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "paletteId": {
            "type": "integer",
            "format": "int32"
          },
          "palette": {
            "$ref": "#/components/schemas/CardPaletteResponse"
          },
          "skipNumber": {
            "type": "integer",
            "format": "int32"
          },
          "hintNumber": {
            "type": "integer",
            "format": "int32"
          },
          "actionType": {
            "type": "string"
          },
          "deckOrder": {
            "type": "integer",
            "format": "int32"
          },
          "updatedUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CardPaletteResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaletteColorResponse"
            }
          }
        }
      },
      "PaletteColorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "hex": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CardDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CardSummaryResponse"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "createdUtc": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CardSaveRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "prompt",
          "actionType"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "prompt": {
            "type": "string",
            "minLength": 1
          },
          "paletteId": {
            "type": "integer",
            "format": "int32"
          },
          "skipNumber": {
            "type": "integer",
            "format": "int32",
            "maximum": 9.0,
            "minimum": 1.0
          },
          "hintNumber": {
            "type": "integer",
            "format": "int32",
            "maximum": 9.0,
            "minimum": 1.0
          },
          "actionType": {
            "type": "string",
            "minLength": 1
          },
          "deckOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}